refactor: extract gateway acp api surface
