+++ crates/daemon/src/gateway/control.rs
@@ -5,7 +5,7 @@ use std::{
 
 use axum::{
     Json, Router,
-    extract::{Query, State},
+    extract::State,
     http::{HeaderMap, StatusCode, header::AUTHORIZATION},
     routing::{get, post},
 };
@@ -13,7 +13,7 @@ use loong_protocol::{
     ControlPlaneConnectErrorCode, ControlPlaneConnectErrorResponse, ControlPlaneConnectRequest,
     ControlPlanePrincipal, ControlPlaneScope,
 };
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use serde_json::{Value, json};
 use tokio::{
     net::TcpListener,
@@ -29,6 +29,9 @@ use super::api_acp::{handle_acp_dispatch, handle_acp_observability, handle_acp_s
 use super::api_events::handle_events;
 use super::api_health::handle_health;
 use super::api_turn::handle_turn;
+use super::acp_api::{
+    handle_gateway_acp_observability, handle_gateway_acp_sessions, handle_gateway_acp_status,
+};
 use super::event_bus::GatewayEventBus;
 use super::lifecycle::{
     combine_gateway_control_task_results, gateway_current_time_ms, gateway_stop_outcome_code,
@@ -50,8 +53,6 @@ use super::pairing_runtime::{
 use super::read_models::{
     GatewayChannelInventoryReadModel, GatewayPairingSessionLeaseReadModel,
     GatewayRuntimeSnapshotReadModel,
-    build_acp_observability_read_model, build_acp_session_list_read_model,
-    build_acp_status_read_model,
     build_gateway_operator_summary_from_registry_read_model,
 };
 use super::state::{
@@ -60,8 +61,8 @@ use super::state::{
 };
 use super::support::{
     build_gateway_channel_inventory_read_model, build_gateway_runtime_snapshot_read_model,
-    gateway_acp_session_list_limit, gateway_control_listener_address_from_port_resolution,
-    resolve_gateway_control_listener_port, serialize_json_value, sort_gateway_acp_sessions,
+    gateway_control_listener_address_from_port_resolution,
+    resolve_gateway_control_listener_port, serialize_json_value,
 };
 
 const GATEWAY_PAIRING_CHALLENGE_MAX_FUTURE_SKEW_MS: u64 = 30_000;
@@ -182,18 +183,6 @@ impl GatewayPairingSessionRequest {
         self.lease = lease;
         Ok(self)
     }
-}
-
-#[derive(Debug, Default, Deserialize)]
-struct GatewayAcpSessionsQuery {
-    limit: Option<usize>,
-}
-
-#[derive(Debug, Default, Deserialize)]
-struct GatewayAcpStatusQuery {
-    session: Option<String>,
-    conversation_id: Option<String>,
-    route_session_id: Option<String>,
 }
 
 #[derive(Clone)]
@@ -668,141 +657,6 @@ async fn handle_gateway_operator_summary(
         pairing_registry.as_ref(),
     );
     gateway_control_payload_response(&summary, "gateway operator summary payload")
-}
-
-async fn handle_gateway_acp_sessions(
-    headers: HeaderMap,
-    State(app_state): State<Arc<GatewayControlAppState>>,
-    Query(query): Query<GatewayAcpSessionsQuery>,
-) -> GatewayControlJsonResponse {
-    let request = match GatewayControlRequest::authorize(&headers, app_state.as_ref()) {
-        Ok(request) => request,
-        Err(response) => return response,
-    };
-    let manager = match request.acp_manager() {
-        Ok(manager) => manager,
-        Err(response) => return response,
-    };
-
-    let sessions_result = manager.list_sessions();
-    let mut sessions = match sessions_result {
-        Ok(sessions) => sessions,
-        Err(error) => {
-            return json_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "acp_sessions_unavailable",
-                error.as_str(),
-            );
-        }
-    };
-
-    sort_gateway_acp_sessions(sessions.as_mut_slice());
-    let matched_count = sessions.len();
-    let limit = gateway_acp_session_list_limit(query.limit);
-    sessions.truncate(limit);
-
-    let payload = build_acp_session_list_read_model(
-        request.app_state().config_path.as_str(),
-        matched_count,
-        sessions.as_slice(),
-    );
-    gateway_control_payload_response(&payload, "gateway ACP sessions payload")
-}
-
-async fn handle_gateway_acp_status(
-    headers: HeaderMap,
-    State(app_state): State<Arc<GatewayControlAppState>>,
-    Query(query): Query<GatewayAcpStatusQuery>,
-) -> GatewayControlJsonResponse {
-    let request = match GatewayControlRequest::authorize(&headers, app_state.as_ref()) {
-        Ok(request) => request,
-        Err(response) => return response,
-    };
-    let config = match request.config() {
-        Ok(config) => config,
-        Err(response) => return response,
-    };
-    let manager = match request.acp_manager() {
-        Ok(manager) => manager,
-        Err(response) => return response,
-    };
-
-    let resolved_session_key = crate::resolve_acp_status_session_key(
-        config,
-        query.session.as_deref(),
-        query.conversation_id.as_deref(),
-        query.route_session_id.as_deref(),
-    );
-    let resolved_session_key = match resolved_session_key {
-        Ok(resolved_session_key) => resolved_session_key,
-        Err(error) if is_gateway_acp_not_found_error(error.as_str()) => {
-            return json_error(StatusCode::NOT_FOUND, "not_found", error.as_str());
-        }
-        Err(error) => {
-            return json_error(StatusCode::BAD_REQUEST, "invalid_selector", error.as_str());
-        }
-    };
-
-    let status_result = manager
-        .get_status(config, resolved_session_key.as_str())
-        .await;
-    let status = match status_result {
-        Ok(status) => status,
-        Err(error) if is_gateway_acp_not_found_error(error.as_str()) => {
-            return json_error(StatusCode::NOT_FOUND, "not_found", error.as_str());
-        }
-        Err(error) => {
-            return json_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "acp_status_unavailable",
-                error.as_str(),
-            );
-        }
-    };
-
-    let payload = build_acp_status_read_model(
-        request.app_state().config_path.as_str(),
-        query.session.as_deref(),
-        query.conversation_id.as_deref(),
-        query.route_session_id.as_deref(),
-        resolved_session_key.as_str(),
-        &status,
-    );
-    gateway_control_payload_response(&payload, "gateway ACP status payload")
-}
-
-async fn handle_gateway_acp_observability(
-    headers: HeaderMap,
-    State(app_state): State<Arc<GatewayControlAppState>>,
-) -> GatewayControlJsonResponse {
-    let request = match GatewayControlRequest::authorize(&headers, app_state.as_ref()) {
-        Ok(request) => request,
-        Err(response) => return response,
-    };
-    let config = match request.config() {
-        Ok(config) => config,
-        Err(response) => return response,
-    };
-    let manager = match request.acp_manager() {
-        Ok(manager) => manager,
-        Err(response) => return response,
-    };
-
-    let snapshot_result = manager.observability_snapshot(config).await;
-    let snapshot = match snapshot_result {
-        Ok(snapshot) => snapshot,
-        Err(error) => {
-            return json_error(
-                StatusCode::INTERNAL_SERVER_ERROR,
-                "acp_observability_unavailable",
-                error.as_str(),
-            );
-        }
-    };
-
-    let payload =
-        build_acp_observability_read_model(request.app_state().config_path.as_str(), &snapshot);
-    gateway_control_payload_response(&payload, "gateway ACP observability payload")
 }
 
 async fn handle_gateway_stop(
